refactor: 달력 임시 리팩토링 진행 및 group by 쿼리 수정
