+++ BE/src/main/resources/application.properties
@@ -31,11 +31,11 @@ spring.thymeleaf.check-template-location=true
 
 
 # profile configuration
-spring.profiles.include=test,API-KEY
+spring.profiles.include=test,API-KEY,AWS-DB
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.url=jdbc:mysql://localhost:3306/refresh?serverTimezone=UTC&characterEncoding=UTF-8
-spring.datasource.username=root
-spring.datasource.password=qwer1234
+#spring.datasource.url=jdbc:mysql://localhost:3306/refresh?serverTimezone=UTC&characterEncoding=UTF-8
+#spring.datasource.username=root
+#spring.datasource.password=qwer1234
