+++ BE/src/main/resources/static/src/pages/calendar/aside/asideHeader/asideHeader.js
@@ -1,4 +1,5 @@
-import { useState, onLoad } from './hook.js';
+import { useState } from '../hook/hook.js';
+import { onLoad } from '../hook/util.js';
 import { getDepartmentNameList } from '../../../../api/calendarApi.js'
